% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-roles.R
\name{init_roles}
\alias{init_roles}
\title{Initialize Variable Roles}
\usage{
init_roles(
  cases_variable = waiver(),
  population_variable = waiver(),
  birth_variable = waiver(),
  death_variable = waiver(),
  series_variable = waiver(),
  median_cases_variable = waiver(),
  count_variable = waiver(),
  total_count_variable = waiver(),
  norm_variable = waiver(),
  time_variable = waiver(),
  period_end_variable = waiver(),
  period_mid_time_variable = waiver(),
  period_mid_date_variable = waiver(),
  period_mid_variable = waiver(),
  period_start_variable = waiver(),
  date_variable = waiver(),
  integer_time_variable = waiver(),
  numeric_time_variable = waiver(),
  time_scale_variable = waiver(),
  period_width_variable = waiver(),
  period_days_variable = waiver(),
  time_group_variable = waiver(),
  time_grouping_variable = waiver(),
  disease_variable = waiver(),
  hierarchical_variable = waiver(),
  nesting_variable = waiver(),
  basal_variable = waiver(),
  title_variable = waiver(),
  grouping_variable = waiver(),
  among_panel_variable = waiver(),
  within_panel_variable = waiver(),
  colour_variable = waiver(),
  category_variable = waiver(),
  categorical_variable = waiver(),
  ...
)
}
\arguments{
\item{cases_variable}{Name of variable containing unstandardized reported incidence.}

\item{population_variable}{Name of variable containing population numbers.}

\item{birth_variable}{Name of variable containing numbers of births.}

\item{death_variable}{Name of variable containing numbers of deaths.}

\item{series_variable}{Name of variable that can be used on the y-axis of a time-series.}

\item{median_cases_variable}{Name of variable containing the median of a set of count variables.}

\item{count_variable}{Name of variable containing a count (e.g., cases, births, deaths, population).}

\item{total_count_variable}{Name of variable containing a marginal total of a set of counts.}

\item{norm_variable}{Name of variable that can be used to normalize another variable (e.g., population normalized cases).}

\item{time_variable}{Name of variable that characterizes the temporal location of the time period.}

\item{period_end_variable}{Name of variable containing}

\item{period_mid_time_variable}{Name of variable containing}

\item{period_mid_date_variable}{Name of variable containing}

\item{period_mid_variable}{Name of variable containing}

\item{period_start_variable}{Name of variable containing}

\item{date_variable}{Name of variable containing}

\item{integer_time_variable}{Name of variable containing}

\item{numeric_time_variable}{Name of variable containing}

\item{time_scale_variable}{Name of variable containing}

\item{period_width_variable}{Name of variable that characterizes the width of the time period.}

\item{period_days_variable}{Name of variable containing}

\item{time_group_variable}{Name of variable containing}

\item{time_grouping_variable}{Name of variable containing}

\item{disease_variable}{Name of variable containing}

\item{hierarchical_variable}{Name of variable containing}

\item{nesting_variable}{Name of variable containing}

\item{basal_variable}{Name of variable containing}

\item{title_variable}{Name of variable containing}

\item{grouping_variable}{Name of variable containing}

\item{among_panel_variable}{Name of variable containing}

\item{within_panel_variable}{Name of variable containing}

\item{colour_variable}{Name of variable containing}

\item{category_variable}{Name of variable containing}

\item{categorical_variable}{Name of variable containing}

\item{...}{Roles not yet thought of.}
}
\description{
Initialize a data structure containing the roles of variables in a data frame
for epidemiological visualization.
}
\details{
Each parameter corresponds to a specific role that a variable can play in the dataset.
The value for each parameter should be the name of the variable (column) in the data frame.
}
