% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizid-class.R
\name{init_vizid}
\alias{init_vizid}
\title{Create a \code{vizid} Object}
\usage{
init_vizid(data, roles = init_roles(), components = init_components())
}
\arguments{
\item{data}{A data frame with infectious disease data.}

\item{roles}{An object produced by \code{\link{init_roles}}, which
maps variable roles to specific names of variables in \code{data} and/or the
return values of data.}

\item{components}{An object produced by \code{init_components}, which
stores \code{ggplot2} components (e.g., the results of functions like
\code{geom_line()} and \code{scales_x_continuous()}) that would be combined into a
plot if the vizid object is passed to the \code{plot()}.}
}
\value{
A data frame of class \code{"vizid"} that knows what roles its variables
will plot in a visualization. These roles are stored in attributes. These
attributes also contains an initially empty list, \code{plot_components}, that
can be filled up with \code{ggplot2} plot components that can be used when the
\code{plot} function is called.
}
\description{
Create a \code{vizid} Object
}
