% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizid-package.R
\docType{package}
\name{vizid-package}
\alias{vizid}
\alias{vizid-package}
\title{vizid: Visualize Infectious Disease Data}
\description{
Tools for visualizing infectious disease data, with special support for datasets that use the data dictionary of the International Infectious Disease Data Archive (IIDDA).
}
\author{
\strong{Maintainer}: Steve walker \email{stevencarlislewalker@gmail.com} (\href{https://orcid.org/0000-0002-4394-9078}{ORCID})

}
\keyword{internal}
